backend/nameforyourprojectbackend/requirements.txt
Django==1.11.3
django-haystack==2.6.0
django-webpack-loader==0.3.3
djangorestframework==3.6.2
gunicorn==19.6.0
numpy==1.13.0
pandas==0.20.2
psycopg2==2.6.2
python-dateutil==2.6.0
pytz==2017.2
requests==2.11.1
six==1.10.0
Whoosh==2.7.4
xlrd==1.0.0
django-cors-headers==2.1.0
firebase-admin==2.8.0